Updated API for keyword movie searches and displaying movies from 2023 to 2024. Revamped movie page to showcase retrieved data and updated results page for keyword search.
I updated the API to allow for movie searches by keyword and to display all movies from the years 2023 to 2024. Additionally, I made changes to the movie page to showcase the movie data retrieved from the API and updated the results page to display all search results corresponding to the provided keyword.